[1/n] Add pre-tokenization for BPE encoding.
GPT-2 uses a relatively complicated regex pattern to do pre-tokenization.
However, Zig only has support for POSIX regex via libc. This means we're
(currently) not able to support pre-tokenization in all cases. For
instance, >=1 whitespace between tokens will not be handled correctly.
We might also not be able to handle unicode, but not sure.